deconstruct train.py
