+++ src/models/data.py
@@ -1,6 +1,26 @@
 from typing import Iterable, Optional
 
 import pandas as pd
+import torch
+from torch.utils.data import DataLoader, Dataset
+
+
+def build_vocab(strings: Iterable[str]):
+    vocab = {"<pad>": 0, "<sos>": 1, "<eos>": 2}
+    idx = 3
+    for sentence in strings:
+        for token in sentence:
+            if token not in vocab:
+                vocab[token] = idx
+                idx += 1
+    return vocab
+
+
+def id2token(vocab: dict[str, int], id: int) -> Optional[str]:
+    for token, idx in vocab.items():
+        if idx == id:
+            return token
+    return None
 
 
 def get_tgt_str(inputs: list[tuple[int]], outputs: list[int]) -> str:
@@ -24,19 +44,46 @@ def get_tgt(row: pd.Series) -> str:
     return get_tgt_str(inputs, outputs)
 
 
-def build_vocab(strings: Iterable[str]):
-    vocab = {"<pad>": 0, "<sos>": 1, "<eos>": 2}
-    idx = 3
-    for sentence in strings:
-        for token in sentence:
-            if token not in vocab:
-                vocab[token] = idx
-                idx += 1
-    return vocab
+def tokenize(string: str, vocab: dict[str, int]) -> list[int]:
+    return (
+        [vocab["<sos>"]] + [vocab[token] for token in string] + [vocab["<eos>"]]
+    )
 
 
-def id2token(vocab: dict[int, str], id: int) -> Optional[str]:
-    for token, idx in vocab.items():
-        if idx == id:
-            return token
-    return None
+class TransformerDataset(Dataset):
+    def __init__(self, df: pd.DataFrame):
+        """
+        Args
+        df: Should have columns "expr", "input" and "output.
+        """
+        self.df = df
+        self.df["tgt_str"] = df["expr"].apply(lambda x: x.replace(" ", ""))
+        self.df["src_str"] = df.apply(get_tgt, axis=1)
+        self.src_vocab = build_vocab(df["src_str"])
+        self.tgt_vocab = build_vocab(df["tgt_str"])
+        self.src_max_len = df["src_str"].apply(len).max()
+        self.tgt_max_len = df["tgt_str"].apply(len).max()
+        self.config = {
+            "src_vocab": self.src_vocab,
+            "tgt_vocab": self.tgt_vocab,
+            "src_max_len": int(self.src_max_len),
+            "tgt_max_len": int(self.tgt_max_len),
+        }
+
+    def __len__(self):
+        return len(self.df)
+
+    def pad_sequence(self, seq, vocab, max_len):
+        return seq + [vocab["<pad>"]] * (max_len - len(seq))
+
+    def __getitem__(self, idx):
+        row = self.df.iloc[idx]
+        src_tokens = tokenize(row["src_str"], self.src_vocab)
+        tgt_tokens = tokenize(row["tgt_str"], self.tgt_vocab)
+        src_padded = self.pad_sequence(
+            src_tokens[: self.src_max_len], self.src_vocab, self.src_max_len
+        )
+        tgt_padded = self.pad_sequence(
+            tgt_tokens[: self.tgt_max_len], self.tgt_vocab, self.tgt_max_len
+        )
+        return torch.tensor(src_padded), torch.tensor(tgt_padded)
